+++ cms/templates/settings.html
@@ -8,7 +8,7 @@
 <%!
   from django.utils.translation import ugettext as _
   from cms.djangoapps.contentstore import utils
-  from openedx.core.djangoapps.certificates.api import can_show_certificate_available_date_field
+  from lms.djangoapps.certificates.api import can_show_certificate_available_date_field
   from openedx.core.djangolib.js_utils import (
       dump_js_escaped_json, js_escaped_string
   )
