+++ lms/djangoapps/certificates/api.py
@@ -9,6 +9,8 @@ certificates models or any other certificates modules.
 
 
 import logging
+from datetime import datetime
+from pytz import UTC
 
 from django.conf import settings
 from django.contrib.auth import get_user_model
@@ -26,6 +28,7 @@ from lms.djangoapps.certificates.generation_handler import (
     generate_certificate_task as _generate_certificate_task,
     is_on_certificate_allowlist as _is_on_certificate_allowlist
 )
+from lms.djangoapps.certificates.config import AUTO_CERTIFICATE_GENERATION as _AUTO_CERTIFICATE_GENERATION
 from lms.djangoapps.certificates.data import CertificateStatuses
 from lms.djangoapps.certificates.models import (
     CertificateAllowlist,
@@ -44,6 +47,7 @@ from lms.djangoapps.certificates.utils import (
     certificate_status as _certificate_status,
     certificate_status_for_student as _certificate_status_for_student,
 )
+from lms.djangoapps.instructor import access
 from openedx.core.djangoapps.content.course_overviews.api import get_course_overview_or_none
 from openedx.core.djangoapps.content.course_overviews.models import CourseOverview
 from xmodule.data import CertificatesDisplayBehaviors
@@ -776,3 +780,83 @@ def certificate_info_for_user(user, course_id, grade, user_is_allowlisted, user_
 def certificate_status_for_student(student, course_id):
     """This returns a dictionary with a key for status, and other information."""
     return _certificate_status_for_student(student, course_id)
+
+def auto_certificate_generation_enabled():
+    return _AUTO_CERTIFICATE_GENERATION.is_enabled()
+
+def _enabled_and_instructor_paced(course):
+    if auto_certificate_generation_enabled():
+        return not course.self_paced
+    return False
+
+def can_show_certificate_available_date_field(course):
+    return _enabled_and_instructor_paced(course)
+
+def can_show_certificate_message(course, student, course_grade, certificates_enabled_for_course):
+    """
+    Returns True if a course certificate message can be shown
+    """
+    auto_cert_gen_enabled = auto_certificate_generation_enabled()
+    has_active_enrollment = CourseEnrollment.is_enrolled(student, course.id)
+    certificates_are_viewable = certificates_viewable_for_course(course)
+    is_beta_tester = access.is_beta_tester(student, course.id)
+    has_passed_or_is_allowlisted = _has_passed_or_is_allowlisted(course, student, course_grade)
+
+    return (
+        (auto_cert_gen_enabled or certificates_enabled_for_course) and
+        has_active_enrollment and
+        certificates_are_viewable and
+        has_passed_or_is_allowlisted and
+        (not is_beta_tester)
+    )
+
+def _course_uses_available_date(course):
+    """Returns if the course has an certificate_available_date set and that it should be used"""
+    if settings.FEATURES.get("ENABLE_V2_CERT_DISPLAY_SETTINGS"):
+        display_behavior_is_valid = course.certificates_display_behavior == CertificatesDisplayBehaviors.END_WITH_DATE
+    else:
+        display_behavior_is_valid = True
+
+    return (
+        can_show_certificate_available_date_field(course)
+        and course.certificate_available_date
+        and display_behavior_is_valid
+    )
+
+def available_date_for_certificate(course, certificate, certificate_available_date=None):
+    """
+    Returns the available date to use with a certificate
+
+    Arguments:
+        course (CourseOverview or course descriptor): The course we're checking
+        certificate (GeneratedCertificate): The certificate we're getting the date for
+        certificate_available_date (datetime): An optional date to override the from the course overview.
+    """
+    if _course_uses_available_date(course):
+        return certificate_available_date or course.certificate_available_date
+    return certificate.modified_date
+
+def display_date_for_certificate(course, certificate):
+    """
+    Returns the display date that a certificate should display.
+
+    Arguments:
+        course (CourseOverview or course descriptor): The course we're getting the date for
+    Returns:
+        datetime.date
+    """
+    if _course_uses_available_date(course) and course.certificate_available_date < datetime.now(UTC):
+        return course.certificate_available_date
+    return certificate.modified_date
+
+def is_valid_pdf_certificate(cert_data):
+    return cert_data.cert_status == CertificateStatuses.downloadable and cert_data.download_url
+
+def _has_passed_or_is_allowlisted(course, student, course_grade):
+    """
+    Returns True if the student has passed this course run, or is on the allowlist for this course run
+    """
+    is_allowlisted = is_on_allowlist(student, course.id)
+    has_passed = course_grade and course_grade.passed
+
+    return has_passed or is_allowlisted
