+++ lms/djangoapps/courseware/date_summary.py
@@ -20,11 +20,10 @@ from lazy import lazy
 from pytz import utc
 
 from common.djangoapps.course_modes.models import CourseMode, get_cosmetic_verified_display_price
-from lms.djangoapps.certificates.api import get_active_web_certificate
+from lms.djangoapps.certificates.api import get_active_web_certificate, can_show_certificate_available_date_field
 from lms.djangoapps.courseware.utils import verified_upgrade_deadline_link, can_show_verified_upgrade
 from lms.djangoapps.verify_student.models import VerificationDeadline
 from lms.djangoapps.verify_student.services import IDVerificationService
-from openedx.core.djangoapps.certificates.api import can_show_certificate_available_date_field
 from openedx.core.djangolib.markup import HTML, Text
 from openedx.features.course_duration_limits.access import get_user_course_expiration_date
 from openedx.features.course_experience import RELATIVE_DATES_FLAG, UPGRADE_DEADLINE_MESSAGE, CourseHomeMessages
