--- openedx/core/djangoapps/certificates/README.rst
@@ -1,1 +1,0 @@
-See ``lms/djangoapps/certificates/README.rst``.
