+++ openedx/core/djangoapps/programs/utils.py
@@ -33,7 +33,6 @@ from openedx.core.djangoapps.catalog.utils import (
     get_fulfillable_course_runs_for_entitlement,
     get_programs,
 )
-from openedx.core.djangoapps.certificates.api import available_date_for_certificate
 from openedx.core.djangoapps.commerce.utils import ecommerce_api_client
 from openedx.core.djangoapps.content.course_overviews.models import CourseOverview
 from openedx.core.djangoapps.credentials.utils import get_credentials
@@ -331,7 +330,10 @@ class ProgramProgressMeter:
                 # Grab the available date and keep it if it's the earliest one for this catalog course.
                 if modes_match and CertificateStatuses.is_passing_status(certificate.status):
                     course_overview = CourseOverview.get_from_id(key)
-                    available_date = available_date_for_certificate(course_overview, certificate)
+                    available_date = certificate_api.available_date_for_certificate(
+                        course_overview,
+                        certificate
+                    )
                     earliest_course_run_date = min(
                         [date for date in [available_date, earliest_course_run_date] if date]
                     )
